refactor(v3.1): rrt 瘦身 58% · 1228 行纯函数迁到 lib/pipeline/score_fns
用户："直接全部开始做吧"

## 搬迁内容 (run_real_test.py → lib/pipeline/score_fns.py)

- `_f` / `score_dimensions` / `generate_panel`
- `_auto_summarize_dim` / `_autofill_qualitative_via_mx`
- `_extract_mx_text` / `generate_synthesis`
- `_is_junk_autofill` / `_AUTOFILL_JUNK_PATTERNS` (v2.12.1)

## 向后兼容

rrt.py 保留 re-export · 所有 `rrt.score_dimensions(...)` 等调用继续工作 ·
pipeline.score / legacy stage1 都不变.

## 行数变化

- run_real_test.py: 2105 → 892 行 (-58%)
- 新 lib/pipeline/score_fns.py: 1271 行

## 测试

- 字符串 grep 式 test 改为同时读 rrt + score_fns 源码 (18 处批量 patch)
- 332 tests 全过
- 真机: 002217 pipeline.score 0.1s · 结果一致

## 下一步 (本分支继续)

- v3.1 · fetcher adapter 内化 · 删 fetch_X.py
- v3.2 · assemble_report.py 瘦身
